KantoCollect/requirements.txt
# =============================================================================
# KANTO COLLECT - Python Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlmodel==0.0.14
sqlalchemy==2.0.25
greenlet>=3.0.0
alembic==1.13.1
asyncpg==0.29.0          # PostgreSQL async driver
aiosqlite==0.19.0        # SQLite async driver (for local dev)

# -----------------------------------------------------------------------------
# Authentication
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# -----------------------------------------------------------------------------
# Validation & Settings
# -----------------------------------------------------------------------------
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# -----------------------------------------------------------------------------
# HTTP Clients (for external APIs)
# -----------------------------------------------------------------------------
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# -----------------------------------------------------------------------------
# Web Scraping
# -----------------------------------------------------------------------------
beautifulsoup4==4.12.3

# -----------------------------------------------------------------------------
# AI / Claude API
# -----------------------------------------------------------------------------
anthropic==0.18.1

# -----------------------------------------------------------------------------
# Image Processing
# -----------------------------------------------------------------------------
pillow==10.2.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
python-dateutil==2.8.2

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0            # For TestClient

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
black==24.1.1
isort==5.13.2
flake8==7.0.0

# -----------------------------------------------------------------------------
# Future: Discord Bot
# -----------------------------------------------------------------------------
# discord.py==2.3.2

# -----------------------------------------------------------------------------
# Future: Shopify
# -----------------------------------------------------------------------------
# ShopifyAPI==12.4.0
